+++ akshare/__init__.py
@@ -3115,9 +3115,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.17.6 fix: fix stock_tfp_em interface
 1.17.7 fix: fix futures_spot_price_daily interface
 1.17.8 fix: fix forex_spot_em interface
+1.17.9 fix: fix stock_zh_ab_comparison_em interface
 """
 
-__version__ = "1.17.8"
+__version__ = "1.17.9"
 __author__ = "AKFamily"
 
 import sys
@@ -4253,6 +4254,7 @@ from akshare.stock_feature.stock_hist_em import (
     stock_sh_a_spot_em,
     stock_sz_a_spot_em,
     stock_zh_b_spot_em,
+    stock_zh_ab_comparison_em,
     stock_zh_a_hist,
     stock_hk_spot_em,
     stock_hk_main_board_spot_em,