+++ docs/changelog.md
@@ -81,6 +81,10 @@
 
 ## 更新说明详情
 
+1.17.9 fix: fix stock_zh_ab_comparison_em interface
+
+	1. 修复 stock_zh_ab_comparison_em 接口
+
 1.17.8 fix: fix forex_spot_em interface
 
 	1. 修复 forex_spot_em 接口
@@ -5617,6 +5621,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.17.9 fix: fix stock_zh_ab_comparison_em interface
 
 1.17.8 fix: fix forex_spot_em interface